Screen: Make OSX OS specific code 64-Bit clean by fixing all compiler warnings
About pointer size mismatches, integer size mismatches etc.
Fixes all compiler warnings, except warnings about deprecated
functionality on 10.7.

+++ PsychSourceGL/Source/Common/Screen/PsychImagingPipelineSupport.c
@@ -3212,7 +3212,7 @@ psych_bool PsychPipelineExecuteBlitter(PsychWindowRecordType *windowRecord, Psyc
 			// On OS-X we can query the OS if the bound shader is running on the GPU or if it is running in emulation mode on the CPU.
 			// This is an expensive operation - it triggers OpenGL internal state revalidation. Only use for debugging and testing!
 			if (PsychPrefStateGet_Verbosity()>5) {
-				long vsgpu=0, fsgpu=0;
+				GLint vsgpu=0, fsgpu=0;
 				CGLGetParameter(CGLGetCurrentContext(), kCGLCPGPUVertexProcessing, &vsgpu);
 				CGLGetParameter(CGLGetCurrentContext(), kCGLCPGPUFragmentProcessing, &fsgpu);
 				printf("PTB-DEBUG: Imaging pipeline GPU shading state: Vertex processing on %s : Fragment processing on %s.\n", (vsgpu) ? "GPU" : "CPU!!", (fsgpu) ? "GPU" : "CPU!!");
